+++ aasrp/discord/direct_message.py
@@ -22,9 +22,6 @@ from django.utils import timezone
 from allianceauth.notifications import notify
 from allianceauth.services.hooks import get_extension_logger
 
-# Alliance Auth (External Libs)
-from app_utils.logging import LoggerAddTag
-
 # AA SRP
 from aasrp import __title__
 from aasrp.app_settings import (
@@ -36,8 +33,9 @@ from aasrp.app_settings import (
     discordproxy_installed,
 )
 from aasrp.constants import DISCORD_EMBED_COLOR_MAP
+from aasrp.providers import AppLogger
 
-logger = LoggerAddTag(get_extension_logger(__name__), __title__)
+logger = AppLogger(get_extension_logger(__name__), __title__)
 
 
 def _aadiscordbot_send_private_message(