+++ aasrp/form.py
@@ -15,9 +15,6 @@ from django.utils.translation import gettext_lazy as _
 # Alliance Auth
 from allianceauth.services.hooks import get_extension_logger
 
-# Alliance Auth (External Libs)
-from app_utils.logging import LoggerAddTag
-
 # AA SRP
 from aasrp import __title__
 from aasrp.constants import KILLBOARD_DATA
@@ -29,6 +26,7 @@ from aasrp.models import (
     SrpRequest,
     UserSetting,
 )
+from aasrp.providers import AppLogger
 
 # Killboard URLs
 zkillboard_base_url: str = KILLBOARD_DATA["zKillboard"][  # pylint: disable=invalid-name
@@ -62,7 +60,7 @@ eve_kill_killmail_url_regex: str = KILLBOARD_DATA[  # pylint: disable=invalid-na
 ]["killmail_url_regex"]
 
 # Initialize a logger with a custom tag for the AA-SRP module
-logger = LoggerAddTag(my_logger=get_extension_logger(__name__), prefix=__title__)
+logger = AppLogger(my_logger=get_extension_logger(__name__), prefix=__title__)
 
 
 def get_mandatory_form_label_text(text: str | Promise) -> str: