+++ aasrp/providers.py
@@ -4,6 +4,9 @@ This module initializes the ESI client with the necessary configuration
 parameters such as compatibility date, user agent name, version, and URL.
 """
 
+# Standard Library
+import logging
+
 # Alliance Auth
 from esi.openapi_clients import ESIClientProvider
 
@@ -31,3 +34,40 @@ esi = ESIClientProvider(
         "GetInsurancePrices",
     ],
 )
+
+
+class AppLogger(logging.LoggerAdapter):
+    """
+    Custom logger adapter that adds a prefix to log messages.
+
+    Taken from the `allianceauth-app-utils` package.
+    Credits to: Erik Kalkoken
+    """
+
+    def __init__(self, my_logger, prefix):
+        """
+        Initializes the AppLogger with a logger and a prefix.
+
+        :param my_logger: Logger instance
+        :type my_logger: logging.Logger
+        :param prefix: Prefix string to add to log messages
+        :type prefix: str
+        """
+
+        super().__init__(my_logger, {})
+
+        self.prefix = prefix
+
+    def process(self, msg, kwargs):
+        """
+        Prepares the log message by adding the prefix.
+
+        :param msg: Log message
+        :type msg: str
+        :param kwargs: Additional keyword arguments
+        :type kwargs: dict
+        :return: Prefixed log message and kwargs
+        :rtype: tuple
+        """
+
+        return f"[{self.prefix}] {msg}", kwargs